Check for winner after each turn
main.go is currently a test of the functionality; will move to actual
unit tests now that the initial "hacking" of the game is done.

Ideally one main.go could become a CLI for playing the game, and another
main.go can be the web interface for an API backend that allows players
to play over a network.